% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool2matrix.R
\name{cool2matrix}
\alias{cool2matrix}
\title{converting cooler format to Hi-C matrices}
\usage{
cool2matrix(file, chr = "chr1")
}
\arguments{
\item{file}{the cooler file to be converted.}

\item{chr}{specifiy which chr to be compared. Default is "chr1".}
}
\value{
a squared HiC matrix that can be recongnized by get.scc.
}
\description{
converting cooler format to Hi-C matrices
}
\examples{
\dontrun{
<<<<<<< HEAD
  # Read a .cool file, extract intrachromosomal interactions
=======
  # Read a .cool file, extract intrachromosomal interactions 
>>>>>>> 5fb05fb8f2e2c0a6bf86ed081614d280cfd0e0b6
  # on chr18 and convert it into a squared matrix
  mat <- cool2matrix(“hic.cool”, chr = ‘chr18’)
}
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a 
stratum-adjusted correlation coefficient. Tao Yang, Feipeng Zhang, Galip
Gurkan Yardimci, Fan Song, Ross C Hardison, William Stafford Noble, 
Feng Yue, Qunhua Li. Genome Research 2017. doi: 10.1101/gr.220640.117
}
