% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.mean.filter.R
\name{fast.mean.filter}
\alias{fast.mean.filter}
\title{A fast mean filter algorithm to smooth a matrix}
\usage{
fast.mean.filter(mat, h)
}
\arguments{
\item{mat}{the matrix to be smoothed.}

\item{h}{window size used to calculate the mean for each cell.}
}
\value{
a smoothed matrix.
}
\description{
A fast mean filter algorithm to smooth a matrix
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a 
stratum-adjusted correlation coefficient. Tao Yang, Feipeng Zhang, Galip
Gurkan Yardimci, Fan Song, Ross C Hardison, William Stafford Noble, 
Feng Yue, Qunhua Li. Genome Research 2017. doi: 10.1101/gr.220640.117
}
