% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.scc.R
\name{get.scc}
\alias{get.scc}
\title{calculate the stratum-adjusted correlation coefficient}
\usage{
get.scc(dat, resol, max)
}
\arguments{
\item{dat}{A matrix of four columns. The first two are the mid-point coordinates of two interacting bin.}

\item{resol}{An integer indicating the resolution of the Hi-C matrix.}

\item{max}{An integer indicating the maximum distance of interaction that is considered.}
}
\value{
A list of results including stratum-specific correlation coefficients, weights, stratum-adjusted correlation
coefficient (scc), and the asymptotic standard deviation of scc.
\itemize{
 \item{corr }{A vector that contains the stratum specific Pearson correlation coefficients.}
 \item{wei }{A vector that contains the weights for each stratum.}
 \item{scc }{Stratum-adjusted correlation coefficients.}
 \item{std }{The asymptotic standard deviation of scc. }
}
}
\description{
calculate the stratum-adjusted correlation coefficient
}
\details{
The function stratifies the Hi-C reads count according to their interacting distance, calculates the Pearson
correlation coefficient for each stratum, then aggregrates them using a weighted average.
}
\examples{
data(HiCR1)
data(HiCR2)
processed <- prep(HiCR1, HiCR2, 1000000, 0, 5000000)

scc.out = get.scc(processed, 1000000, 5000000)
scc.out$scc
scc.out$std
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a stratum-adjusted correlation coefficient. 
Tao Yang, Feipeng Zhang, Galip Gurkan Yardimci, Ross C Hardison, William Stafford Noble, Feng Yue, Qunhua Li. 
bioRxiv 101386; doi: https://doi.org/10.1101/101386.
}

