% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic2mat.R
\name{hic2mat}
\alias{hic2mat}
\title{extract HiC data from .hic file and convert it to squared HiC matrix}
\usage{
hic2mat(file = "path/to/file", chromosome1, chromosome2, resol,
  method = "NONE")
}
\arguments{
\item{file}{the path to the .hic file.}

\item{chromosome1}{character number which specify chromosome to extract.}

\item{chromosome2}{for intra-chromosome data, set it same as chromosome 1; 
for inter-chromosome, set to another chromosome.}

\item{resol}{resolution, i.e., bin size.}

\item{method}{specifies what data to extract, raw or normalized. 
Must be one of "NONE", "VC", "VC_SQRT", "KR". "NONE" will gives raw counts. 
VC is vanilla coverage, VC_SQRT is square root of vanilla coverage, and KR 
is Knight-Ruiz or Balanced normalization.}
}
\value{
a squared HiC matrix that can be recongnized by get.scc.
}
\description{
extract HiC data from .hic file and convert it to squared HiC matrix
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a 
stratum-adjusted correlation coefficient. Tao Yang, Feipeng Zhang, Galip
Gurkan Yardimci, Fan Song, Ross C Hardison, William Stafford Noble, 
Feng Yue, Qunhua Li. Genome Research 2017. doi: 10.1101/gr.220640.117
}
