% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCR2.R
\docType{data}
\name{HiCR2}
\alias{HiCR2}
\title{HiCR2}
\format{An object of class \code{data.frame} with 1283 rows and 1286 columns.}
\usage{
HiCR2
}
\description{
A Hi-C matrix in the format of \eqn{N*(3+N)}, the additional first three columns contains the chromosome
information and coordinates of interaction bins. First is chromosome name, the second and third are the mid-points of
the contacting bins.
}
\references{
Evaluating the reproducibility of Hi-C data. Tao Yang, Feng Yue, Qunhua Li. 2016.
}
\keyword{datasets}

