% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htrain.R
\name{htrain}
\alias{htrain}
\title{Train the smoothing parameter (neighborhood size).}
\usage{
htrain(R1, R2, resol, max, range)
}
\arguments{
\item{R1}{A Hi-C intra-chromosome matrix.}

\item{R2}{The other intra-chromosome matrix to compare with.}

\item{resol}{An integer indicating the resolution of the Hi-C matrix.}

\item{max}{An integer indicating the maximum distance of interaction that is considered.}

\item{range}{A vector of integers that are used to search the optimal smoothing parameter(i.g., 0:10).}
}
\description{
Train the smoothing parameter (neighborhood size).
}
\details{
A fraction (10\%) of data are first randomly sampled, then the scc for the sampled data is
computed at a series of smoothing parameterts in the ascending order. The samllest h at which the
increment of scc is less than 0.01 is saved. This procedure is repeat 10 times, and the mode of the
10 \code{h}'s is outputed as the estimated optimal neighborhood size.
}
\examples{
data(HiCR1)
data(HiCR2)
h_hat = htrain(HiCR1, HiCR2, 40000, 5000000, 0:2)
}
\references{
Evaluating the reproducibility of Hi-C data. Tao Yang, Feng Yue, Qunhua Li. 2016.
}

