% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2mat.R
\name{bed2mat}
\alias{bed2mat}
\title{converting 3-column contact matrix to squared HiC matrix}
\usage{
bed2mat(bed, resol = "NONE")
}
\arguments{
\item{bed}{the 3-column contact matrix; first and second columns are genome coordinates or bin index.}

\item{resol}{resolution in base pair unit. Set it to "NONE" if the first two columns are bin indices.}
}
\value{
a squared HiC matrix that can be recongnized by get.scc.
}
\description{
converting 3-column contact matrix to squared HiC matrix
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a 
stratum-adjusted correlation coefficient. Tao Yang, Feipeng Zhang, Galip
Gurkan Yardimci, Fan Song, Ross C Hardison, William Stafford Noble, 
Feng Yue, Qunhua Li. Genome Research 2017. doi: 10.1101/gr.220640.117
}
