% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCR2.R
\docType{data}
\name{HiCR2}
\alias{HiCR2}
\title{HiCR2}
\format{An object of class \code{data.frame} with 52 rows and 55 columns.}
\usage{
HiCR2
}
\description{
A Hi-C matrix in the format of \eqn{N*(3+N)}, 
the additional first three columns contains the chromosome
information and coordinates of interaction bins. First is 
chromosome name, the second and third are the mid-points of
the contacting bins.
}
\references{
HiCRep: assessing the reproducibility of Hi-C 
data using a stratum-adjusted correlation coefficient. 
Tao Yang, Feipeng Zhang, Galip Gurkan Yardimci, Ross C Hardison,
William Stafford Noble, Feng Yue, Qunhua Li. 
bioRxiv 101386; doi: https://doi.org/10.1101/101386.
}
\keyword{datasets}

