% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCR2.R
\docType{data}
\name{HiCR2}
\alias{HiCR2}
\title{HiCR2}
\format{An object of class \code{data.frame} with 52 rows and 52 columns.}
\usage{
HiCR2
}
\description{
A Hi-C matrix in the format of \eqn{N*N}.
}
\references{
HiCRep: assessing the reproducibility of Hi-C data using a 
stratum-adjusted correlation coefficient. Tao Yang, Feipeng Zhang, Galip
Gurkan Yardimci, Fan Song, Ross C Hardison, William Stafford Noble, 
Feng Yue, Qunhua Li. Genome Research 2017. doi: 10.1101/gr.220640.117
}
\keyword{datasets}
